+++ backend/src/bork-protocol/utils/tweet-analysis/process-knowledge.ts
